Refactor: Split user tests for each API into separate files(WIP)
