+++ tests/api/room/room_member/test_invite_to_room_test.py
@@ -1,15 +1,10 @@
-import pytest
 import grpc
 from unittest.mock import MagicMock
-from typing import Generator, Tuple
-from grpc import ServicerContext
-from pytest_mock import MockerFixture
+from typing import Tuple
+
 
 
 from src.api.room.room_member_service import RoomMemberServicer
-from src.api.room.room_member_service import RoomMembers
-from src.proto_generated.nori.v0.room.room_user_request_pb2 import RoomUserRequest
-from src.utils.token_helper import generate_jwt_token
 from src.proto_generated.nori.v0.room.room_id_pb2 import RoomId
 from src.proto_generated.nori.v0.user.user_id_pb2 import UserId
 
@@ -17,32 +12,12 @@ from src.proto_generated.nori.v0.room.member.invite_user_to_room_request_pb2 imp
     InviteUserToRoomRequest,
 )
 from google.protobuf.empty_pb2 import Empty
+from mock_repo import mock_repositories, grpc_context
 
 
 
-@pytest.fixture
-def mock_repositories(
-    mocker: MockerFixture,
-) -> Generator[Tuple[MagicMock, MagicMock, MagicMock], None, None]:
-    """Mock UserRepository, RoomRepository, RoomMemberRepository"""
-    mock_db_session = MagicMock()
-    mocker.patch("src.api.room.room_member_service.get_db", return_value=mock_db_session)
 
-    mock_user_repo: MagicMock = mocker.patch("src.api.room.room_member_service.UserRepo")
-    mock_room_repo: MagicMock = mocker.patch("src.api.room.room_member_service.RoomRepo")
-    mock_room_member_repo: MagicMock = mocker.patch(
-        "src.api.room.room_member_service.RoomMemberRepo"
-    )
 
-    yield mock_user_repo, mock_room_repo, mock_room_member_repo
-
-@pytest.fixture
-def grpc_context() -> MagicMock:
-    """Mock gRPC context with valid authorization metadata"""
-    context: MagicMock = MagicMock(spec=ServicerContext)
-    token = generate_jwt_token(subject="test")
-    context.invocation_metadata.return_value = (("authorization", token),)
-    return context
 
 def test_invite_to_room_room_not_found(
     mock_repositories: Tuple[MagicMock, MagicMock, MagicMock], grpc_context: MagicMock
@@ -121,56 +96,3 @@ def test_invite_to_room_success(
         1, [3, 4]
     )
     assert isinstance(response, Empty)
-
-
-def test_join_room_success(
-    mock_repositories: Tuple[MagicMock, MagicMock, MagicMock], grpc_context: MagicMock
-) -> None:
-    mock_user_repo, mock_room_repo, mock_room_member_repo = mock_repositories
-    room_service = RoomMemberServicer()
-
-    mock_room_repo.return_value.exists_room.return_value = True
-    mock_user_repo.return_value.exists_user.return_value = True
-
-    request = RoomUserRequest(room_id=RoomId(id=123), user_id=UserId(id=1))
-    response = room_service.JoinRoom(request, grpc_context)
-
-    mock_room_member_repo.return_value.create_room_member.assert_called_once_with(
-        RoomMembers(room_id=123, user_id=1)
-    )
-
-    assert isinstance(response, Empty)
-
-
-def test_join_room_room_not_found(
-    mock_repositories: Tuple[MagicMock, MagicMock, MagicMock], grpc_context: MagicMock
-) -> None:
-    _, mock_room_repo, _ = mock_repositories
-    room_service = RoomMemberServicer()
-
-    mock_room_repo.return_value.exists_room.return_value = False
-
-    request = RoomUserRequest(room_id=RoomId(id=999), user_id=UserId(id=1))
-    response = room_service.JoinRoom(request, grpc_context)
-
-    grpc_context.set_code.assert_called_once_with(grpc.StatusCode.NOT_FOUND)
-    grpc_context.set_details.assert_called_once_with("Room with ID 999 not found.")
-    assert isinstance(response, Empty)
-
-
-def test_join_room_user_not_found(
-    mock_repositories: Tuple[MagicMock, MagicMock, MagicMock], grpc_context: MagicMock
-) -> None:
-    mock_user_repo, mock_room_repo, _ = mock_repositories
-    room_service = RoomMemberServicer()
-
-    mock_room_repo.return_value.exists_room.return_value = True
-    mock_user_repo.return_value.exists_user.return_value = False
-
-    request = RoomUserRequest(room_id=RoomId(id=123), user_id=UserId(id=1))
-    response = room_service.JoinRoom(request, grpc_context)
-
-    grpc_context.set_code.assert_called_once_with(grpc.StatusCode.NOT_FOUND)
-    grpc_context.set_details.assert_called_once_with("User with ID 1 not found.")
-    assert isinstance(response, Empty)
-
